Support non-primary, non-foreign-key prefixed IDs
== Use Case

This is needed when storing IDs generated outside of the application,
or which may not have a canonical reference within the application's
database.

In those cases, the prefixed UUID is neither a primary key, nor does
it reference via e.g. `belong_to` any other schema in the application.

== Solution

Instead of checking for the field being a primary key, instead check
for it being a foreign key. In the case of it not being a foreign key,
the prefix is immediately known, so store only that along with the Uniq
UUID configuration.

== Testing

Unit tests were added to demonstrate both the bug before this change,
and how this change addresses the need describedin Use Case.